refactor: `LottoCompany`에게 로또 추첨의 책임 분배
